+++ lib/nemo_ui.dart
@@ -6,6 +6,7 @@ export 'src/components/nemo_surface.dart';
 export 'src/components/nemo_switch.dart';
 export 'src/foundation/nemo_asset_resolver.dart';
 export 'src/foundation/nemo_localizations.dart';
+export 'src/foundation/nemo_material.dart';
 export 'src/foundation/nemo_motion.dart';
 export 'src/foundation/nemo_theme.dart';
 export 'src/foundation/nemo_theme_data.dart';